partial restructure
Some work on restructuring the evaluator
